[refactor] decide moved to src (#69)
decide function moved to src for better folder structure, and fixed imports in test
