+++ src/LAUNCH.py
@@ -1,4 +1,4 @@
-from src.utils import InvalidInputException
+from utils import InvalidInputException
 
 def LAUNCH(FUV):
     """
